Use logger/profiler from node-core
