+++ packages/node-core/src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './object';
 export * from './promise';
 export * from './graphql';
 export * from './sync-helper';
+export * from './batch-size';
