+++ packages/node/src/configure/configure.module.ts
@@ -6,10 +6,9 @@ import path from 'path';
 import { DynamicModule, Global, Module } from '@nestjs/common';
 import { getProjectRootAndManifest, IPFS_REGEX } from '@subql/common';
 import { SubstrateProjectNetworkConfig } from '@subql/common-substrate';
+import { getYargsOption, getLogger, setLevel } from '@subql/node-core';
 import { IConfig, MinConfig, NodeConfig } from '@subql/node-core/configure';
 import { camelCase, last, omitBy, isNil } from 'lodash';
-import { getLogger, setLevel } from '../utils/logger';
-import { getYargsOption } from '@subql/node-core';
 import { SubqueryProject } from './SubqueryProject';
 const logger = getLogger('configure');
