+++ packages/node/src/indexer/api.service.ts
@@ -7,10 +7,10 @@ import { ApiPromise, WsProvider } from '@polkadot/api';
 import { ApiOptions, RpcMethodResult } from '@polkadot/api/types';
 import { RuntimeVersion } from '@polkadot/types/interfaces';
 import { AnyFunction, DefinitionRpcExt } from '@polkadot/types/types';
+import { getLogger } from '@subql/node-core';
 import { IndexerEvent, NetworkMetadataPayload } from '@subql/node-core/events';
 import { SubstrateBlock } from '@subql/types';
 import { SubqueryProject } from '../configure/SubqueryProject';
-import { getLogger } from '../utils/logger';
 import { ApiAt } from './types';
 import { HttpProvider } from './x-provider/http';
