+++ packages/node/src/indexer/dynamic-ds.service.ts
@@ -4,11 +4,11 @@
 import assert from 'assert';
 import { Injectable } from '@nestjs/common';
 import { isCustomDs, isRuntimeDs } from '@subql/common-substrate';
+import { getLogger } from '@subql/node-core';
+import { MetadataRepo } from '@subql/node-core/indexer/entities';
 import { Transaction } from 'sequelize/types';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
-import { getLogger } from '../utils/logger';
 import { DsProcessorService } from './ds-processor.service';
-import { MetadataRepo } from '@subql/node-core/indexer/entities';
 
 const logger = getLogger('dynamic-ds');
