+++ packages/node/src/indexer/sandbox.service.ts
@@ -7,7 +7,7 @@ import {
   isDatasourceV0_2_0,
   SubstrateDataSource,
 } from '@subql/common-substrate';
-import { getYargsOption } from '@subql/node-core';
+import { getYargsOption, getLogger } from '@subql/node-core';
 import { NodeConfig } from '@subql/node-core/configure';
 import { StoreService } from '@subql/node-core/indexer';
 import { timeout } from '@subql/node-core/utils';
@@ -16,7 +16,6 @@ import { levelFilter } from '@subql/utils';
 import { merge } from 'lodash';
 import { NodeVM, NodeVMOptions, VMScript } from 'vm2';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
-import { getLogger } from '../utils/logger';
 import { getProjectEntry } from '../utils/project';
 import { ApiService } from './api.service';
 import { ApiAt } from './types';
