+++ packages/node/src/indexer/worker/block-dispatcher.service.ts
@@ -10,14 +10,12 @@ import { RuntimeVersion } from '@polkadot/types/interfaces';
 import { hexToU8a, u8aEq } from '@polkadot/util';
 import { NodeConfig } from '@subql/node-core/configure';
 import { IndexerEvent } from '@subql/node-core/events';
+import { getLogger } from '@subql/node-core/logger';
 import { SubstrateBlock } from '@subql/types';
 import chalk from 'chalk';
 import { last } from 'lodash';
 import { AutoQueue, Queue } from '../../utils/autoQueue';
-import { getLogger } from '../../utils/logger';
-import { profilerWrap } from '../../utils/profiler';
-import * as SubstrateUtil from '../../utils/substrate';
-import { getYargsOption } from '../../yargs';
+import * as SubstrateUtil  from '../../utils/substrate';
 import { ApiService } from '../api.service';
 import { IndexerManager } from '../indexer.manager';
 import { ProjectService } from '../project.service';
