+++ packages/node/src/indexer/worker/worker.service.ts
@@ -4,9 +4,9 @@
 import { threadId } from 'node:worker_threads';
 import { Injectable } from '@nestjs/common';
 import { RuntimeVersion } from '@polkadot/types/interfaces';
+import { getLogger } from '@subql/node-core/logger';
 import { NodeConfig } from '@subql/node-core/configure';
 import { AutoQueue } from '../../utils/autoQueue';
-import { getLogger } from '../../utils/logger';
 import { fetchBlocksBatches } from '../../utils/substrate';
 import { ApiService } from '../api.service';
 import { IndexerManager } from '../indexer.manager';
