+++ packages/node/src/indexer/worker/worker.ts
@@ -5,7 +5,7 @@ import assert from 'assert';
 import { threadId } from 'node:worker_threads';
 import { INestApplication } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-import { getLogger, NestLogger } from '../../utils/logger';
+import { getLogger, NestLogger } from '@subql/node-core/logger';
 import { IndexerManager } from '../indexer.manager';
 import { registerWorker } from './worker.builder';
 import { WorkerModule } from './worker.module';
